+++ ddb-client/src/test/java/org/opensearch/remote/metadata/client/impl/DDBOpenSearchClientTests.java
@@ -87,12 +87,14 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import static org.opensearch.core.xcontent.XContentParserUtils.ensureExpectedToken;
+import static org.opensearch.remote.metadata.client.AbstractSdkClient.DEFAULT_TENANT;
 import static org.opensearch.remote.metadata.client.impl.DDBOpenSearchClient.simulateOpenSearchResponse;
 import static org.opensearch.remote.metadata.common.CommonValue.REMOTE_METADATA_GLOBAL_RESOURCE_CACHE_TTL_KEY;
 import static org.opensearch.remote.metadata.common.CommonValue.REMOTE_METADATA_GLOBAL_TENANT_ID_KEY;
 import static org.opensearch.remote.metadata.common.CommonValue.TENANT_ID_FIELD_KEY;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -802,7 +804,12 @@ public class DDBOpenSearchClientTests {
         when(dynamoDbAsyncClient.updateItem(updateItemRequestArgumentCaptor.capture())).thenReturn(
             CompletableFuture.completedFuture(UpdateItemResponse.builder().build())
         );
-        SdkClient sdkClient = SdkClientFactory.wrapSdkClientDelegate(
+        SdkClient multiTenancyNotEnabledClient = SdkClientFactory.wrapSdkClientDelegate(
+            new DDBOpenSearchClient(dynamoDbAsyncClient, aosOpenSearchClient, Map.of()),
+            false,
+            null
+        );
+        SdkClient globalResourceEnabledClient = SdkClientFactory.wrapSdkClientDelegate(
             new DDBOpenSearchClient(
                 dynamoDbAsyncClient,
                 aosOpenSearchClient,
@@ -818,17 +825,53 @@ public class DDBOpenSearchClientTests {
             true,
             GLOBAL_TENANT_ID
         );
-        UpdateDataObjectResponse updateResponse = sdkClient.updateDataObjectAsync(updateRequest, testThreadPool.executor(TEST_THREAD_POOL))
-            .toCompletableFuture()
-            .join();
-        assertEquals(TEST_ID, updateResponse.id());
-        UpdateItemRequest updateItemRequest = updateItemRequestArgumentCaptor.getValue();
+        // multi-tenancy not enabled case, the update is successful.
+        UpdateDataObjectResponse multiTenancyNotEnabledResponse = multiTenancyNotEnabledClient.updateDataObjectAsync(
+            updateRequest,
+            testThreadPool.executor(TEST_THREAD_POOL)
+        ).toCompletableFuture().join();
+        assertEquals(TEST_ID, multiTenancyNotEnabledResponse.id());
+        UpdateItemRequest updateItemRequest1 = updateItemRequestArgumentCaptor.getValue();
         assertEquals(TEST_ID, updateRequest.id());
-        assertEquals(TEST_INDEX, updateItemRequest.tableName());
-        assertEquals(TEST_ID, updateItemRequest.key().get(RANGE_KEY).s());
-        assertEquals(TENANT_ID, updateItemRequest.key().get(HASH_KEY).s());
-        assertEquals("foo", updateItemRequest.expressionAttributeValues().get(":source").m().get("data").s());
-        assertEquals("old_value", updateItemRequest.expressionAttributeValues().get(":source").m().get("old_key").s());
+        assertEquals(TEST_INDEX, updateItemRequest1.tableName());
+        assertEquals(TEST_ID, updateItemRequest1.key().get(RANGE_KEY).s());
+        assertEquals(TENANT_ID, updateItemRequest1.key().get(HASH_KEY).s());
+        assertEquals("foo", updateItemRequest1.expressionAttributeValues().get(":source").m().get("data").s());
+        assertEquals("old_value", updateItemRequest1.expressionAttributeValues().get(":source").m().get("old_key").s());
+
+        // multi-tenancy not enabled and tenant id null case, the update is successful and tenant id is default tenant id.
+        UpdateDataObjectRequest tenantIdNullRequest = UpdateDataObjectRequest.builder()
+            .id(TEST_ID)
+            .index(TEST_INDEX)
+            .retryOnConflict(1)
+            .dataObject(testDataObject)
+            .build();
+        UpdateDataObjectResponse multiTenancyNotEnabledTenantNullResponse = multiTenancyNotEnabledClient.updateDataObjectAsync(
+            tenantIdNullRequest,
+            testThreadPool.executor(TEST_THREAD_POOL)
+        ).toCompletableFuture().join();
+        assertEquals(TEST_ID, multiTenancyNotEnabledTenantNullResponse.id());
+        UpdateItemRequest updateItemRequest2 = updateItemRequestArgumentCaptor.getValue();
+        assertEquals(TEST_ID, tenantIdNullRequest.id());
+        assertEquals(TEST_INDEX, updateItemRequest2.tableName());
+        assertEquals(TEST_ID, updateItemRequest2.key().get(RANGE_KEY).s());
+        assertEquals(DEFAULT_TENANT, updateItemRequest2.key().get(HASH_KEY).s());
+        assertEquals("foo", updateItemRequest2.expressionAttributeValues().get(":source").m().get("data").s());
+        assertEquals("old_value", updateItemRequest2.expressionAttributeValues().get(":source").m().get("old_key").s());
+
+        // global resource enabled case, the update is also successful.
+        UpdateDataObjectResponse globalResourceEnabledResponse = globalResourceEnabledClient.updateDataObjectAsync(
+            updateRequest,
+            testThreadPool.executor(TEST_THREAD_POOL)
+        ).toCompletableFuture().join();
+        assertEquals(TEST_ID, globalResourceEnabledResponse.id());
+        UpdateItemRequest updateItemRequest3 = updateItemRequestArgumentCaptor.getValue();
+        assertEquals(TEST_ID, tenantIdNullRequest.id());
+        assertEquals(TEST_INDEX, updateItemRequest3.tableName());
+        assertEquals(TEST_ID, updateItemRequest3.key().get(RANGE_KEY).s());
+        assertEquals(TENANT_ID, updateItemRequest3.key().get(HASH_KEY).s());
+        assertEquals("foo", updateItemRequest3.expressionAttributeValues().get(":source").m().get("data").s());
+        assertEquals("old_value", updateItemRequest3.expressionAttributeValues().get(":source").m().get("old_key").s());
     }
 
     @Test
@@ -877,6 +920,42 @@ public class DDBOpenSearchClientTests {
         UpdateResponse response = UpdateResponse.fromXContent(updateResponse.parser());
         assertEquals(5, response.getSeqNo());
 
+    }
+
+    @Test
+    public void testUpdateDataObjectAsync_getItemException_failToUpdate() {
+        UpdateDataObjectRequest updateRequest = UpdateDataObjectRequest.builder()
+            .id(TEST_ID)
+            .index(TEST_INDEX)
+            .tenantId(TENANT_ID)
+            .retryOnConflict(1)
+            .dataObject(testDataObject)
+            .build();
+        when(dynamoDbAsyncClient.getItem(any(GetItemRequest.class))).thenReturn(
+            CompletableFuture.failedFuture(new RuntimeException("Failed to get item"))
+        );
+        SdkClient sdkClient = SdkClientFactory.wrapSdkClientDelegate(
+            new DDBOpenSearchClient(
+                dynamoDbAsyncClient,
+                aosOpenSearchClient,
+                Map.of(
+                    TENANT_ID_FIELD_KEY,
+                    TEST_TENANT_ID,
+                    REMOTE_METADATA_GLOBAL_TENANT_ID_KEY,
+                    GLOBAL_TENANT_ID,
+                    REMOTE_METADATA_GLOBAL_RESOURCE_CACHE_TTL_KEY,
+                    TEST_GLOBAL_RESOURCE_CACHE_TTL
+                )
+            ),
+            true,
+            GLOBAL_TENANT_ID
+        );
+        CompletionException exception = assertThrows(
+            CompletionException.class,
+            () -> sdkClient.updateDataObjectAsync(updateRequest, testThreadPool.executor(TEST_THREAD_POOL)).toCompletableFuture().join()
+        );
+        assertInstanceOf(RuntimeException.class, exception);
+        assertEquals("java.lang.RuntimeException: Failed to get item", exception.getMessage());
     }
 
     @Test