+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Enhancements
 Add CMK support to accept CMK to encrypt/decrypt customer data. ([#271](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/271))
 ### Bug Fixes
+* Fix error when updating model status ([#291](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/291))
 ### Infrastructure
 ### Documentation
 ### Maintenance